requirements.txt
gevent==21.8.0
locust_plugins==2.1.1
locustio==0.999
